first version of dragnn wrapper
